♻️ refactor(git): move git info logic into local-file-shell (one impl, CLI-ready)
Extract the working-directory git read logic into `@lobechat/local-file-shell`
(shared by desktop + CLI), so there's one implementation behind the desktop
display, the device `gitInfo` RPC, and anything the CLI builds.

- New `local-file-shell/src/git`: repoType helpers + getGitBranch /
  getLinkedPullRequest / getGitWorkingTreeStatus / getGitAheadBehind / gitInfo
  aggregate + the shared `DeviceGitInfo` type (pure node, moved verbatim).
- Desktop `GitCtr` git read methods now delegate to it; `gitInfo` is also an
  `@IpcMethod` now. `utils/git` re-exports the helpers (SystemCtr unaffected).

The CLI already depends on `@lobechat/local-file-shell`, so it can call `gitInfo`
directly — a headless `lh connect` device gaining git status is now just wiring.
