+++ apps/desktop/src/main/controllers/GitCtr.ts
@@ -22,42 +22,19 @@ import type {
   GitWorkingTreeStatus,
   SubmoduleWorkingTreePatches,
 } from '@lobechat/electron-client-ipc';
+import {
+  type DeviceGitInfo,
+  getGitAheadBehind as computeGitAheadBehind,
+  getGitBranch as computeGitBranch,
+  getGitWorkingTreeStatus as computeGitWorkingTreeStatus,
+  getLinkedPullRequest as computeLinkedPullRequest,
+  gitInfo as computeGitInfo,
+} from '@lobechat/local-file-shell';
 
-import { detectRepoType, resolveGitDir } from '@/utils/git';
+import { detectRepoType } from '@/utils/git';
 import { createLogger } from '@/utils/logger';
 
 import { ControllerModule, IpcMethod } from './index';
-
-/**
- * Local mirror of `@lobechat/types`' `DeviceGitInfo` — the desktop tsconfig only
- * sees a stub of `@lobechat/types` that doesn't include it, so inline the shape
- * (same pattern as `BuiltinServerRuntimeOutput` in GatewayConnectionCtr). Keep in
- * sync with the canonical type in `packages/types/src/agentExecution/index.ts`.
- */
-interface DeviceGitInfo {
-  aheadBehind: {
-    ahead: number;
-    behind: number;
-    hasUpstream: boolean;
-    pushTarget?: string;
-    pushTargetExists?: boolean;
-    upstream?: string;
-  };
-  info: {
-    branch?: string;
-    detached?: boolean;
-    extraCount?: number;
-    ghMissing?: boolean;
-    pullRequest?: { number: number; state: string; title: string; url: string } | null;
-  };
-  workingStatus: {
-    added: number;
-    clean: boolean;
-    deleted: number;
-    modified: number;
-    total: number;
-  };
-}
 
 const logger = createLogger('controllers:GitCtr');
 
@@ -481,54 +458,17 @@ export default class GitController extends ControllerModule {
    */
   @IpcMethod()
   async getGitBranch(dirPath: string): Promise<GitBranchInfo> {
-    try {
-      const gitDir = await resolveGitDir(dirPath);
-      if (!gitDir) return {};
-
-      const head = (await readFile(path.join(gitDir, 'HEAD'), 'utf8')).trim();
-      const refMatch = /^ref:\s*refs\/heads\/(.+)$/.exec(head);
-      if (refMatch) {
-        return { branch: refMatch[1] };
-      }
-      // Detached HEAD — HEAD file contains the full sha
-      if (/^[\da-f]{40}$/i.test(head)) {
-        return { branch: head.slice(0, 7), detached: true };
-      }
-      return {};
-    } catch {
-      return {};
-    }
+    return computeGitBranch(dirPath);
   }
 
   /**
-   * Aggregate git status for a working directory into one payload, served over
-   * the device RPC so a remote device (or web client) can render branch / file
-   * changes / PR. Mirrors the renderer's `useGitInfo` + `useWorkingTreeStatus` +
-   * `useGitAheadBehind`. Returns empty-ish data for a non-git directory.
+   * Aggregate git status (branch + linked PR + working tree + ahead/behind) for a
+   * directory. The single entry point shared by the local desktop display, the
+   * device `gitInfo` RPC, and the CLI — implemented in `@lobechat/local-file-shell`.
    */
+  @IpcMethod()
   async gitInfo(params: { isGithub?: boolean; scope: string }): Promise<DeviceGitInfo> {
-    const dirPath = params.scope;
-    const { branch, detached } = await this.getGitBranch(dirPath);
-
-    let info: DeviceGitInfo['info'] = { branch, detached };
-    // PR lookup only for a real branch on a GitHub remote (mirrors useGitInfo).
-    if (branch && !detached && params.isGithub) {
-      const pr = await this.getLinkedPullRequest({ branch, path: dirPath });
-      info = {
-        branch,
-        detached,
-        extraCount: pr.extraCount,
-        ghMissing: pr.status === 'gh-missing',
-        pullRequest: pr.pullRequest,
-      };
-    }
-
-    const [workingStatus, aheadBehind] = await Promise.all([
-      this.getGitWorkingTreeStatus(dirPath),
-      this.getGitAheadBehind(dirPath),
-    ]);
-
-    return { aheadBehind, info, workingStatus };
+    return computeGitInfo(params);
   }
 
   /**
@@ -541,58 +481,7 @@ export default class GitController extends ControllerModule {
     branch: string;
     path: string;
   }): Promise<GitLinkedPullRequestResult> {
-    const { path: dirPath, branch } = payload;
-    if (!branch) {
-      return { pullRequest: null, status: 'ok' };
-    }
-
-    const execFileAsync = promisify(execFile);
-    try {
-      const { stdout } = await execFileAsync(
-        'gh',
-        [
-          'pr',
-          'list',
-          '--head',
-          branch,
-          '--state',
-          'open',
-          '--limit',
-          '5',
-          '--json',
-          'number,url,title,state',
-        ],
-        { cwd: dirPath, timeout: 8000 },
-      );
-      const parsed = JSON.parse(stdout.trim() || '[]') as Array<{
-        number: number;
-        state: string;
-        title: string;
-        url: string;
-      }>;
-      if (parsed.length === 0) {
-        return { pullRequest: null, status: 'ok' };
-      }
-      const [primary, ...rest] = parsed;
-      return {
-        extraCount: rest.length,
-        pullRequest: primary,
-        status: 'ok',
-      };
-    } catch (error: any) {
-      const code = error?.code;
-      const stderr: string = error?.stderr ?? '';
-      // `gh` binary not on PATH
-      if (code === 'ENOENT') {
-        return { pullRequest: null, status: 'gh-missing' };
-      }
-      // gh reports auth issues via stderr; treat as a soft-fail
-      if (/auth\s+login|not\s+logged\s+in|authentication/i.test(stderr)) {
-        return { pullRequest: null, status: 'gh-missing' };
-      }
-      logger.debug('[getLinkedPullRequest] failed', { branch, code, stderr });
-      return { pullRequest: null, status: 'error' };
-    }
+    return computeLinkedPullRequest(payload);
   }
 
   /**
@@ -697,42 +586,7 @@ export default class GitController extends ControllerModule {
    */
   @IpcMethod()
   async getGitWorkingTreeStatus(dirPath: string): Promise<GitWorkingTreeStatus> {
-    const execFileAsync = promisify(execFile);
-    try {
-      const { stdout } = await execFileAsync('git', ['status', '--porcelain', '-u', '-z'], {
-        cwd: dirPath,
-        timeout: 5000,
-      });
-      const tokens = stdout.split('\0');
-      let added = 0;
-      let modified = 0;
-      let deleted = 0;
-      let i = 0;
-      while (i < tokens.length) {
-        const entry = tokens[i];
-        i++;
-        if (entry.length < 2) continue;
-        const x = entry[0];
-        const y = entry[1];
-        // R/C entries carry an extra source-path token we must consume.
-        if (x === 'R' || x === 'C') i++;
-        if (x === '?' && y === '?') {
-          added++;
-        } else if (x === '!' && y === '!') {
-          // ignored — skip
-        } else if (x === 'D' || y === 'D') {
-          deleted++;
-        } else if (x === 'A' || y === 'A') {
-          added++;
-        } else {
-          modified++;
-        }
-      }
-      const total = added + modified + deleted;
-      return { added, clean: total === 0, deleted, modified, total };
-    } catch {
-      return { added: 0, clean: true, deleted: 0, modified: 0, total: 0 };
-    }
+    return computeGitWorkingTreeStatus(dirPath);
   }
 
   /**
@@ -1195,66 +1049,7 @@ export default class GitController extends ControllerModule {
    */
   @IpcMethod()
   async getGitAheadBehind(dirPath: string): Promise<GitAheadBehind> {
-    const execFileAsync = promisify(execFile);
-    try {
-      await execFileAsync('git', ['fetch', '--no-tags', '--quiet', 'origin'], {
-        cwd: dirPath,
-        timeout: 10_000,
-      });
-    } catch {
-      // swallow — fall through to compute against cached refs
-    }
-    try {
-      const { stdout: upstreamOut } = await execFileAsync(
-        'git',
-        ['rev-parse', '--abbrev-ref', '--symbolic-full-name', '@{u}'],
-        { cwd: dirPath, timeout: 5000 },
-      );
-      const upstream = upstreamOut.trim();
-      if (!upstream) return { ahead: 0, behind: 0, hasUpstream: false };
-
-      const { stdout } = await execFileAsync(
-        'git',
-        ['rev-list', '--left-right', '--count', `${upstream}...HEAD`],
-        { cwd: dirPath, timeout: 5000 },
-      );
-      const [behindStr, aheadStr] = stdout.trim().split(/\s+/);
-      const behind = Number.parseInt(behindStr ?? '0', 10) || 0;
-      const ahead = Number.parseInt(aheadStr ?? '0', 10) || 0;
-
-      // `git push -u origin HEAD` always targets origin/<current-branch-name>,
-      // which may differ from upstream (the branched-off-canary case).
-      let pushTarget: string | undefined;
-      let pushTargetExists = false;
-      try {
-        const { stdout: branchOut } = await execFileAsync(
-          'git',
-          ['symbolic-ref', '--short', 'HEAD'],
-          { cwd: dirPath, timeout: 5000 },
-        );
-        const branch = branchOut.trim();
-        if (branch) {
-          pushTarget = `origin/${branch}`;
-          try {
-            await execFileAsync(
-              'git',
-              ['rev-parse', '--verify', '--quiet', `refs/remotes/${pushTarget}`],
-              { cwd: dirPath, timeout: 5000 },
-            );
-            pushTargetExists = true;
-          } catch {
-            pushTargetExists = false;
-          }
-        }
-      } catch {
-        // detached HEAD — leave pushTarget undefined
-      }
-
-      return { ahead, behind, hasUpstream: true, pushTarget, pushTargetExists, upstream };
-    } catch {
-      // No upstream configured, detached HEAD, or git error — all treated as "no upstream"
-      return { ahead: 0, behind: 0, hasUpstream: false };
-    }
+    return computeGitAheadBehind(dirPath);
   }
 
   /**
