+++ packages/local-file-shell/src/index.ts
@@ -1,6 +1,7 @@
 export * from './contentSearch';
 export * from './file';
 export * from './fileSearch';
+export * from './git';
 export { createLogger, type Logger, type LoggerFactory, setLoggerFactory } from './logger';
 export * from './shell';
 export type { ToolCategory, ToolDetector } from './toolDetector';
